requirements.txt
flask>=3.0.0
mediapipe>=0.10.9
opencv-python>=4.9.0
numpy>=1.26.0
Pillow>=10.0.0
tensorflow>=2.15.0
